chore(actor): Move `ActorTuple` to its own file so that it's not tied to models during import
We want to use `ActorTuple` in the issue platform, but it causes a lot of import loops currently. Moving this to a separate place so it can more easily be imported
