+++ tests/sentry/deletions/test_organization.py
@@ -2,7 +2,6 @@ from uuid import uuid4
 
 from sentry.discover.models import DiscoverSavedQuery, DiscoverSavedQueryProject
 from sentry.incidents.models.alert_rule import AlertRule, AlertRuleStatus
-from sentry.models.actor import ActorTuple
 from sentry.models.commit import Commit
 from sentry.models.commitauthor import CommitAuthor
 from sentry.models.dashboard import Dashboard
@@ -30,6 +29,7 @@ from sentry.testutils.cases import TransactionTestCase
 from sentry.testutils.hybrid_cloud import HybridCloudTestMixin
 from sentry.testutils.outbox import outbox_runner
 from sentry.testutils.silo import assume_test_silo_mode
+from sentry.utils.actor import ActorTuple
 
 
 class DeleteOrganizationTest(TransactionTestCase, HybridCloudTestMixin):
